Update README.md
This is our FY Project from MMCOE, IT branch on course registration system. The problem statement of our project was to develop a comprehensive course registration system for students to enroll in classes, view schedules, manage courses, and handle administrative tasks like grading and attendance. This system will be used by students and faculty.